hiding from search results
multiple pages from the tutorial are showing up under key search terms and crowding out the product documentation and tutorials intended for customer audiences. Since the primary purpose of this tutorial is for select internal, partner, and POC customers with access to specially environments, it is not ideal content for random searchers. I am hiding this content for now as a test. if there is a need to make this content searchable, we must come up with a strategy for titles and descriptions so that it is clear from the ExL SERP what this content is.